backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-telegram-bot==20.7
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
bcrypt==4.2.0
